webhook/requirements.txt
Flask==2.2.2
Flask-DebugToolbar==0.13.1
Jinja2==3.1.2
Werkzeug==2.2.2
pymongo==4.3.3
python-dotenv==0.21.1